Carve `TestRaggedTensors` out of the monolith (wala/ML#635).
The 66 `testRagged*` methods move verbatim into the new class; the
monolith keeps 974, preserving the 1,040 total.
